Don't remove duplicated descriptions
It's not a lot of data anyway. For /cats/57?lang=tw,
The first form and evolved form are sharing the same description,
and the true form is using a different one. If we remove the
duplicated descriptions, the true form description will be mistaken
as the evolved form, which is wrong.