Fix Hermit Cat max DPS area
We make area_range always a range so any mix of attacks we can
calculate max DPS area easily. However, we do still want to display
simplified range, because that's easier to understand and generally
the default inner range (starting range) isn't exactly important
because it's the default and should be taken as granted if not
specified.

This should make sure it displays as before, yet internally it should
calculate the correct numbers.